backend/app/requirements.txt
Flask-Cors
Flask-RESTful
PyJWT